% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRAPI.R
\name{SCRAPI}
\alias{SCRAPI}
\title{SCRAPI v2.2: The smolt companion to SCOBI}
\usage{
SCRAPI(
  smoltData = NULL,
  Dat = "CollectionDate",
  Rr = "Rear",
  Primary = "GenStock",
  Secondary = NA,
  passageData = NULL,
  strat = "Week",
  dat = "SampleEndDate",
  tally = "SampleCount",
  samrate = "SampleRate",
  guidance = "GuidanceEfficiency",
  collaps = "Collapse",
  Run = "output",
  RTYPE = "W",
  REARSTRAT = TRUE,
  alph = 0.1,
  B = 5000,
  dateFormat = "\%m/\%d/\%Y"
)
}
\arguments{
\item{smoltData}{the data.frame (or path to .csv) containing biological data for
smolts to be analyzed. Should contain all smolts trapped in a given migratory year
and species (sthd, ch0, or ch1). Use \code{lgr2SCRAPI()} to format raw data from
LGTrappingDB.}

\item{Dat}{column in \code{smoltData} containing sample date for each smolt}

\item{Rr}{column in \code{smoltData} containing rear type (W or HNC)}

\item{Primary}{primary category in \code{smoltData} to estimate (e.g. GenStock)}

\item{Secondary}{secondary category to estimate nested within Primary; use
\code{NA} if no secondary decomposition is desired}

\item{passageData}{data.frame (or path to .csv) containing smolt passage data with
calendar week, sampling rate, daily smolt count, guidance efficiency, and
collapsing scheme}

\item{strat}{column in \code{passageData} containing calendar weeks}

\item{dat}{column in \code{passageData} containing sample dates}

\item{tally}{column in \code{passageData} containing daily smolt count in the trap}

\item{samrate}{column in \code{passageData} containing daily trap sample rate}

\item{guidance}{column in \code{passageData} containing daily guidance efficiency}

\item{collaps}{column in \code{passageData} containing the collapsing scheme}

\item{Run}{label used as prefix for all output files}

\item{RTYPE}{rear type to analyse: "W" (wild) or "HNC" (hatchery no-clip)}

\item{REARSTRAT}{if \code{TRUE}, proportion wild is calculated on a
time-stratified basis; if \code{FALSE} it is pooled across the emigration}

\item{alph}{alpha for confidence intervals (e.g. 0.10 gives 90\% CIs)}

\item{B}{number of bootstrap iterations}

\item{dateFormat}{date format string used in fish and passage data
(default \code{"\%m/\%d/\%Y"})}
}
\value{
NULL (results written to CSV files and printed to console)
}
\description{
Perform compositional analyses of smolts at Lower Granite Dam.
}
\examples{
\dontrun{
SCRAPI(smoltData = sthdScrapiInput, Primary = "GenStock", Secondary = "fwAge",
       passageData = sthdSmoltPassData, Run = "sthdSmoltTest",
       RTYPE = "W", alph = 0.1, B = 200)
}

}
\author{
Kirk Steinhorst and Mike Ackerman
}
