% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoltEASE-package.R
\docType{package}
\name{smoltEASE-package}
\alias{smoltEASE}
\alias{smoltEASE-package}
\title{smoltEASE package imports}
\description{
Preserve the package's existing imports when regenerating NAMESPACE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dsmi313/smoltEASE}
  \item Report bugs at \url{https://github.com/dsmi313/smoltEASE/issues}
}

}
\author{
\strong{Maintainer}: David Smith \email{david.smith@idfg.idaho.gov}

Other contributors:
\itemize{
  \item Thomas Delomas [contributor]
}

}
\keyword{internal}
